+++ demo/src/i18n/client.ts
@@ -1,4 +1,4 @@
 import { createI18nClient } from "@fransek/next-i18n/lib/core/client";
-import i18n from "./i18n";
+import i18nConfig from "./i18nConfig";
 
-export const { useContent, useLocale } = createI18nClient(i18n);
+export const { useContent, useLocale } = createI18nClient(i18nConfig);
