+++ demo/src/i18n/i18nConfig.ts
