+++ demo/src/i18n/server.ts
@@ -1,5 +1,5 @@
 import { createI18nServerClient } from "@fransek/next-i18n/lib/core/server";
-import i18n from "./i18n";
+import i18nConfig from "./i18nConfig";
 
 export const { getContent, getLocale, middleware } =
-  createI18nServerClient(i18n);
+  createI18nServerClient(i18nConfig);
